+++ benchmarks/scaling/TODO.md
@@ -15,8 +15,9 @@
     - [x] Implement method to load some trains
   - [ ] Generate the signal (i.e. raw data)
     - [x] Add gaussian noise to the signal
-    - [ ] Add template waveforms to the signal
+    - [ ] ~~Add template waveforms to the signal~~
     - [ ] Use the `synthetic_generator` and `writer` block to generate the signal
+  - [x] Move code from `io.pregenerate` to `net.pregenerator`
 - [ ] Move the utils from `utils.py` to `circusort`'s core
 - [ ] Explain how to use a non-empty initial template dictionary (in the README)
 - [x] Correct README (bugs Pierre)
