+++ circusort/__init__.py
@@ -1,6 +1,7 @@
 from . import block
 from . import cli
 from . import io
+from . import net
 
 from .base import create_director
 from .base import utils
