+++ circusort/io/__init__.py
@@ -4,6 +4,7 @@ from . import configure
 from . import generate
 from .load import load, load_peaks, load_spikes, load_times
 from .configure import Configuration
-from .probe import generate_probe, save_probe, load_probe
 from .parse import parse_parameters
-from .pregenerate import pregenerate
+from .probe import generate_probe, save_probe, load_probe
+from .trains import generate_trains, save_trains, load_trains
+from .template import generate_templates, save_templates, load_templates
